# apihub-server/src/main/java/com/apihub/agent/model/dto/UserSwitchRequest.java
package com.apihub.agent.model.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

public class UserSwitchRequest {

    @NotNull
    @Positive
    private Long userId;

    public Long getUserId() {
        return userId;
    }

    public void setUserId(Long userId) {
        this.userId = userId;
    }
}